feat: optimized seo
