+++ src/app/contact/page.tsx
@@ -1,14 +1,39 @@
 // src/app/contact/page.tsx
 import type { Metadata } from 'next';
-import ContactPageClient from '@/components/ContactPageClient'; // Import the new client component
+import ContactPageClient from '@/components/ContactPageClient';
+import { faqData } from '@/data/faqData'; // Import the FAQ data
 
-// This is a Server Component, so we can export metadata
 export const metadata: Metadata = {
   title: 'Contact & Book a Consultation | Toasted Sesame Therapy',
   description: 'Ready to start? Reach out to book a free, no-pressure 15-minute consultation. Your journey toward healing is one conversation away.'
 };
 
-// This is the actual page component now. It just renders the client part.
 export default function ContactPage() {
-  return <ContactPageClient />;
+  // Generate the JSON for the schema
+  const faqSchema = {
+    "@context": "https://schema.org",
+    "@type": "FAQPage",
+    "mainEntity": faqData.map(item => ({
+      "@type": "Question",
+      "name": item.question,
+      "acceptedAnswer": {
+        "@type": "Answer",
+        // This removes any HTML like <br /> for the schema script
+        "text": item.answer.replace(/<[^>]*>?/gm, ' ')
+      }
+    }))
+  };
+
+  return (
+    <>
+      {/* Add the schema script here */}
+      <script
+        type="application/ld+json"
+        dangerouslySetInnerHTML={{ __html: JSON.stringify(faqSchema) }}
+      />
+
+      {/* Render the client component */}
+      <ContactPageClient />
+    </>
+  );
 }
