+++ src/app/layout.tsx
@@ -21,8 +21,42 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  // Define your schema objects here
+  const websiteSchema = {
+    "@context": "https://schema.org",
+    "@type": "WebSite",
+    "name": "Toasted Sesame Therapy",
+    "url": "https://toastedsesametherapy.com", // Remember to replace this with your actual domain
+  };
+
+  const medicalBusinessSchema = {
+    "@context": "https://schema.org",
+    "@type": "MedicalBusiness",
+    "@id": "https://toastedsesametherapy.com/#medicalbusiness", // Remember to replace this
+    "name": "Toasted Sesame Therapy",
+    "address": {
+      "@type": "PostalAddress",
+      "addressLocality": "Atlanta",
+      "addressRegion": "GA",
+      "addressCountry": "US"
+    },
+    "url": "https://toastedsesametherapy.com", // Remember to replace this
+    "logo": "https://pvbdrbaquwivhylsmagn.supabase.co/storage/v1/object/public/tst-assets/logo/TST-LOGO-WHITE.svg"
+  };
+
   return (
     <html lang="en">
+      <head>
+        {/* Add the schema scripts to the head */}
+        <script
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{ __html: JSON.stringify(websiteSchema) }}
+        />
+        <script
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{ __html: JSON.stringify(medicalBusinessSchema) }}
+        />
+      </head>
       <body
         className={`${workSans.variable} font-sans bg-tst-cream text-black antialiased`}
       >
